gfxaccel: structural remediation
Compositor contract: metal_compositor.h now documents production reality (Present is the sole presenter; SubmitFrame is a validated overlay mailbox). RAVE/GL overlays render ping-pong so Present can never sample a texture an engine is redrawing; GL clears the cached overlay on unbind; dead DSp underlay route removed.

Blockers: indexed-palette latch is dirty-gated with back-buffer sync (was strobing stale/fresh every VBL); NQD transparent(36)/ hilite(50) pen compares fixed (byte-order) at 8/16/32bpp.

Pacing + GPU gates: deadline-based per-engine frame pacing with satisfied-boundary semantics (co-resident DSp+GL engines pace to the same vblank — fixes structural half-rate); RAVE ring gated by per-slot completion semaphores with near-exhaustion mid-frame flush; GL triple ring gated; ring shutdown refcounted; texture updates fenced via staging+blit; max-frame-rate throttle implemented.

Heap lifecycle: release-before-reset enforced (live-count gate + GPU-completion latch); eviction restores purgeable state and runs from both present paths; DSp heap no longer ratchets (alt-buffer backings release on background, rebuild from guest staging on foreground); Resize validates-then-swaps with rollback.

GL: CFM double-decode fixed across dispatch (FPR wire format, GPRshadow-skip for mixed signatures, 13-FPR cap — gluLookAt upZ); ARB_multitexture, EXT_blend_color/equation wired end-to-end; full GLU bridge (54 exports incl. gluBuild2DMipmaps); fog_coord de-advertised, secondary_color wired; honest GL_VERSION; synthetic TVECT self-patch guarded; tessellation counts clamped.

RAVE: modifiedRect guest-memory writeback removed; color clears preserve depth; pass continuations store/load attachments correctly (incl. MSAA); AccessZBuffer commits in-flight work, resolves MSAA depth, returns big-endian floats; pipeline cache keyed on blend+mask; UV2 streams ring-staged past 4KB; RTT textures survive either-order delete.

DrawSprocket: busyProc polarity/ABI/timeout per the 1.7 API docs (timeout proceeds, never aborts); Switch routes through SetState with failure unwind; Reserve colorTable applied (with guest-RAM extent validation); front-staging publish keyed on content; background/foreground lifecycle drains synchronously (bitmask pending flags).

QuickDraw (NQD): overlapping same-surface blits ordered (CPU scratch for standard Boolean, decline-to-software for packed/arithmetic); region masks decode pixel columns at 16/32bpp; sub-byte origins declined to the software path; legacy entries validate extents; dispatcher CPU fallbacks flush the GPU batch first.

Threading: main==emul ground truth documented (gfxaccel_threading_policy.h); VBL fan-out guarded against nested re-entry; SwapBuffers revalidates across pump windows; DMC veto binding preserved and rollback compensated. New policy headers: gfx_color_policy.h (single gamma owner: compositor LUT applies guest gamma + 1.8-to-sRGB once, on every slot incl. overlays), gfx_frame_pacing_policy.h, overlay_clear_policy.h

+++ SheepShaver/src/gfxaccel/include/gl_engine.h
@@ -747,12 +747,12 @@ enum {
  *
  *  float_mask: bit N = 1 means arg N is a float/double (comes from FPR).
  *              bit N = 0 means arg N is int/pointer (comes from GPR).
- *  For functions with > 8 args, only the first 8 are tracked (PPC ABI
- *  passes additional args on the stack regardless of type).
+ *  PPC ABI passes floating-point args in FPR1-FPR13, so signatures can
+ *  describe up to 13 PPC-visible floating-point args.
  */
 struct GLFuncSignature {
     uint8_t num_args;    // Total argument count (excluding ctx pointer)
-    uint8_t float_mask;  // Bit per arg: 1=float/double, 0=int/ptr
+    uint16_t float_mask; // Bit per arg: 1=float/double, 0=int/ptr
 };
